tests: split PatternFamilyTests.swift (1,917 lines → 5 files)
Follow-on to #490.  71 tests in a single XCTestCase was one parallel
worker.  Each subclass now exercises one concern.

Layout:
  PatternFamilyTestCase.swift        — base + family fixtures + plumbing
  PatternFamilyRendererTests.swift   — JSONValue, renderer, default-args,
                                       family-vars, validation (26)
  PatternFamilyApplyTests.swift      — applyPatternFamilies + family deps
                                       + family points + authored order (18)
  PatternFamilySectionsTests.swift   — v0.4.96 section grouping (5)
  PatternFamilyKindsTests.swift      — approxEquality, variableEquality,
                                       stdoutEquality renderer + validation (22)

Behaviour-preserving refactor; 71 tests pass under --parallel.
